Move xorshift to separate, package-visible class
